fix: move test_ollama_integration back to tests/ to prevent CI failures and duplicates
